backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
rembg==2.0.50
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Production / manufacturing pipeline
reportlab>=4.0.0
shapely>=2.0.0
svgwrite>=1.4.0
requests>=2.31.0
supabase>=2.0.0
# pypotrace>=0.3.0  # optional: for raster->vector trace; or use opencv-python
opencv-python-headless>=4.8.0
